+++ projeto/tinder-evolution/src/main/java/br/com/cwi/tinderevolution/dominio/Serie.java
@@ -8,12 +8,12 @@ public class Serie {
     public String nome;
     public String diretor;
     public LocalDate dataLancamento;
-    public Double numeroDeTemporadas;
-    public Double numeroDeEpisodios;
+    public int numeroDeTemporadas;
+    public int numeroDeEpisodios;
     public CategoriasSerie categoria;
     public String sinopse;
 
-    public Serie(String nome, String diretor, LocalDate dataLancamento, Double numeroDeTemporadas, Double numeroDeEpisodios, CategoriasSerie categoria, String sinopse) {
+    public Serie(String nome, String diretor, LocalDate dataLancamento, int numeroDeTemporadas, int numeroDeEpisodios, CategoriasSerie categoria, String sinopse) {
         this.nome = nome;
         this.diretor = diretor;
         this.dataLancamento = dataLancamento;
@@ -22,6 +22,7 @@ public class Serie {
         this.categoria = categoria;
         this.sinopse = sinopse;
     }
+
 
     @Override
     public String toString() {
